разюивает код на модули
